fix(backend): исправлена уязвимость, позволяющая обновлять и удалять пользователя от лица другого

+++ backend/app/api/routes/users.py
@@ -1,8 +1,9 @@
 from uuid import UUID
 
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException, status
 
-from api.deps import get_user_service
+from api.deps import get_current_user, get_user_service
+from core.models import User
 from core.schemas import UserCreate, UserRead, UserUpdate
 from services import UserService
 from services.exceptions import UserNotFoundError
@@ -44,14 +45,20 @@ async def get_user(
 async def update_user(
     user_id: UUID,
     update_schema: UserUpdate,
+    current_user: User = Depends(get_current_user),
     service: UserService = Depends(get_user_service),
 ) -> UserRead:
+    if current_user.id != user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Forbidden")
     return await service.update_user(user_id, update_schema)
 
 
 @router.delete("/{user_id}", status_code=204)
 async def delete_user(
     user_id: UUID,
+    current_user: User = Depends(get_current_user),
     service: UserService = Depends(get_user_service),
 ) -> None:
+    if current_user.id != user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Forbidden")
     await service.delete_user(user_id)
